test(ai): split openrouter provider spec into focused files
Break the 1000+ line monolith into constructor, generate, streaming, token ceiling, and error handling specs so each contract surface stays under 800 lines.
