+++ tests/unit/ai/providers/openrouter.streaming.spec.ts
@@ -1,15 +1,7 @@
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
 
-import {
-  DEFAULT_OUTPUT_TOKEN_CEILING,
-  getOutputTokenCeiling,
-} from '@/features/ai/cost';
 import { AI_DEFAULT_MODEL } from '@/features/ai/ai-models';
-import { ProviderInvalidResponseError } from '@/features/ai/providers/errors';
-import {
-  OpenRouterProvider,
-  type OpenRouterProviderConfig,
-} from '@/features/ai/providers/openrouter';
+import { OpenRouterProvider } from '@/features/ai/providers/openrouter';
 import type { StreamEventLike } from '@/features/ai/providers/openrouter-response';
 import {
   collectOpenRouterStream,
@@ -29,356 +21,7 @@ describe('OpenRouterProvider', () => {
     vi.unstubAllEnvs();
   });
 
-  describe('constructor', () => {
-    it('throws error when model is not provided', () => {
-      const { client } = createOpenRouterMockClient();
-      expect(
-        () => new OpenRouterProvider({} as OpenRouterProviderConfig, client),
-      ).toThrow('OpenRouterProvider requires a model to be specified');
-    });
-
-    it('throws error when API key is not provided', () => {
-      vi.stubEnv('OPENROUTER_API_KEY', '');
-      expect(
-        () => new OpenRouterProvider({ model: OPENROUTER_TEST_MODEL }),
-      ).toThrow('OPENROUTER_API_KEY is not set');
-    });
-
-    it('uses custom model when specified', async () => {
-      const { client, send } = createOpenRouterMockClient();
-      send.mockResolvedValueOnce({
-        model: 'anthropic/claude-3-sonnet',
-        choices: [
-          {
-            message: {
-              content: JSON.stringify(VALID_PLAN_RESPONSE),
-            },
-          },
-        ],
-      });
-
-      const provider = new OpenRouterProvider(
-        { model: 'anthropic/claude-3-sonnet' },
-        client,
-      );
-
-      const result = await provider.generate(OPENROUTER_SAMPLE_INPUT);
-
-      expect(send).toHaveBeenCalledWith(
-        expect.objectContaining({
-          model: 'anthropic/claude-3-sonnet',
-        }),
-        expect.any(Object),
-      );
-      expect(result.metadata.model).toBe('anthropic/claude-3-sonnet');
-    });
-
-    it('sends fallback routes using models when configured', async () => {
-      const { client, send } = createOpenRouterMockClient();
-      send.mockResolvedValueOnce({
-        model: AI_DEFAULT_MODEL,
-        choices: [
-          {
-            message: {
-              content: JSON.stringify(VALID_PLAN_RESPONSE),
-            },
-          },
-        ],
-        usage: {
-          promptTokens: 20,
-          completionTokens: 40,
-          totalTokens: 60,
-        },
-      });
-
-      const provider = new OpenRouterProvider(
-        {
-          model: 'anthropic/claude-haiku-4.5',
-          fallbackModels: [AI_DEFAULT_MODEL],
-        },
-        client,
-      );
-      const result = await provider.generate(OPENROUTER_SAMPLE_INPUT);
-      await collectOpenRouterStream(result.stream);
-
-      const [request] = send.mock.calls[0] ?? [];
-      expect(request).toMatchObject({
-        models: ['anthropic/claude-haiku-4.5', AI_DEFAULT_MODEL],
-      });
-      expect(request).not.toHaveProperty('model');
-      expect(result.metadata.model).toBe(AI_DEFAULT_MODEL);
-    });
-  });
-
   describe('generate', () => {
-    it('generates a valid plan from string content response', async () => {
-      const { client, send } = createOpenRouterMockClient();
-      send.mockResolvedValueOnce({
-        choices: [
-          {
-            message: {
-              content: JSON.stringify(VALID_PLAN_RESPONSE),
-            },
-          },
-        ],
-        usage: {
-          promptTokens: 100,
-          completionTokens: 500,
-          totalTokens: 600,
-        },
-      });
-
-      const provider = new OpenRouterProvider(
-        { model: OPENROUTER_TEST_MODEL },
-        client,
-      );
-      const result = await provider.generate(OPENROUTER_SAMPLE_INPUT);
-
-      const rawText = await collectOpenRouterStream(result.stream);
-      const parsed = JSON.parse(rawText);
-
-      expect(parsed.modules).toHaveLength(3);
-      expect(parsed.modules[0].title).toBe('Introduction to TypeScript');
-      expect(result.metadata.provider).toBe('openrouter');
-      expect(result.metadata.model).toBe(OPENROUTER_TEST_MODEL);
-    });
-
-    it('generates a valid plan from array content response', async () => {
-      const { client, send } = createOpenRouterMockClient();
-      send.mockResolvedValueOnce({
-        choices: [
-          {
-            message: {
-              content: [
-                { type: 'text', text: JSON.stringify(VALID_PLAN_RESPONSE) },
-              ],
-            },
-          },
-        ],
-        usage: {
-          promptTokens: 100,
-          completionTokens: 500,
-          totalTokens: 600,
-        },
-      });
-
-      const provider = new OpenRouterProvider(
-        { model: OPENROUTER_TEST_MODEL },
-        client,
-      );
-      const result = await provider.generate(OPENROUTER_SAMPLE_INPUT);
-
-      const rawText = await collectOpenRouterStream(result.stream);
-      const parsed = JSON.parse(rawText);
-
-      expect(parsed.modules).toHaveLength(3);
-    });
-
-    it('returns correct usage metadata', async () => {
-      const { client, send } = createOpenRouterMockClient();
-      send.mockResolvedValueOnce({
-        choices: [
-          {
-            message: {
-              content: JSON.stringify(VALID_PLAN_RESPONSE),
-            },
-          },
-        ],
-        usage: {
-          promptTokens: 150,
-          completionTokens: 750,
-          totalTokens: 900,
-        },
-      });
-
-      const provider = new OpenRouterProvider(
-        { model: OPENROUTER_TEST_MODEL },
-        client,
-      );
-      const result = await provider.generate(OPENROUTER_SAMPLE_INPUT);
-
-      expect(result.metadata.usage).toEqual({
-        promptTokens: 150,
-        completionTokens: 750,
-        totalTokens: 900,
-        providerReportedCostUsd: undefined,
-      });
-    });
-
-    it('maps non-streaming usage.cost (USD) to providerReportedCostUsd', async () => {
-      const { client, send } = createOpenRouterMockClient();
-      send.mockResolvedValueOnce({
-        choices: [
-          {
-            message: {
-              content: JSON.stringify(VALID_PLAN_RESPONSE),
-            },
-          },
-        ],
-        usage: {
-          promptTokens: 10,
-          completionTokens: 20,
-          totalTokens: 30,
-          cost: 0.004567,
-        },
-      });
-
-      const provider = new OpenRouterProvider(
-        { model: OPENROUTER_TEST_MODEL },
-        client,
-      );
-      const result = await provider.generate(OPENROUTER_SAMPLE_INPUT);
-
-      expect(result.metadata.usage?.providerReportedCostUsd).toBe(0.004567);
-    });
-
-    it('handles missing usage data gracefully', async () => {
-      const { client, send } = createOpenRouterMockClient();
-      send.mockResolvedValueOnce({
-        choices: [
-          {
-            message: {
-              content: JSON.stringify(VALID_PLAN_RESPONSE),
-            },
-          },
-        ],
-      });
-
-      const provider = new OpenRouterProvider(
-        { model: OPENROUTER_TEST_MODEL },
-        client,
-      );
-      const result = await provider.generate(OPENROUTER_SAMPLE_INPUT);
-
-      // Missing usage data is reported as undefined (not silently defaulted to 0)
-      // so downstream normalization can detect and alert on it.
-      expect(result.metadata.usage).toEqual({
-        promptTokens: undefined,
-        completionTokens: undefined,
-        totalTokens: undefined,
-        providerReportedCostUsd: undefined,
-      });
-    });
-
-    it('calls SDK with correct parameters including maxTokens', async () => {
-      const { client, send } = createOpenRouterMockClient();
-      send.mockResolvedValueOnce({
-        choices: [
-          {
-            message: {
-              content: JSON.stringify(VALID_PLAN_RESPONSE),
-            },
-          },
-        ],
-      });
-
-      const provider = new OpenRouterProvider(
-        {
-          model: 'anthropic/claude-3-opus',
-          temperature: 0.7,
-        },
-        client,
-      );
-
-      await provider.generate(OPENROUTER_SAMPLE_INPUT);
-
-      expect(send).toHaveBeenCalledWith(
-        expect.objectContaining({
-          model: 'anthropic/claude-3-opus',
-          temperature: 0.7,
-          stream: true,
-          responseFormat: { type: 'json_object' },
-          maxTokens: expect.any(Number),
-          messages: expect.arrayContaining([
-            expect.objectContaining({ role: 'system' }),
-            expect.objectContaining({ role: 'user' }),
-          ]),
-        }),
-        expect.any(Object),
-      );
-    });
-
-    it('uses the safest output ceiling across the route models', async () => {
-      const { client, send } = createOpenRouterMockClient();
-      send.mockResolvedValueOnce({
-        model: AI_DEFAULT_MODEL,
-        choices: [
-          {
-            message: {
-              content: JSON.stringify(VALID_PLAN_RESPONSE),
-            },
-          },
-        ],
-      });
-
-      const provider = new OpenRouterProvider(
-        {
-          model: 'openai/gpt-4o',
-          fallbackModels: [AI_DEFAULT_MODEL],
-        },
-        client,
-      );
-      await provider.generate(OPENROUTER_SAMPLE_INPUT);
-
-      const requestBody = send.mock.calls[0][0] as {
-        maxTokens: number;
-      };
-      expect(requestBody.maxTokens).toBe(DEFAULT_OUTPUT_TOKEN_CEILING);
-    });
-
-    it('includes topic in user prompt', async () => {
-      const { client, send } = createOpenRouterMockClient();
-      send.mockResolvedValueOnce({
-        choices: [
-          {
-            message: {
-              content: JSON.stringify(VALID_PLAN_RESPONSE),
-            },
-          },
-        ],
-      });
-
-      const provider = new OpenRouterProvider(
-        { model: OPENROUTER_TEST_MODEL },
-        client,
-      );
-      await provider.generate(OPENROUTER_SAMPLE_INPUT);
-
-      const callArgs = send.mock.calls[0][0];
-      const userMessage = callArgs.messages.find(
-        (m: { role: string }) => m.role === 'user',
-      );
-
-      expect(userMessage.content).toContain('TypeScript Fundamentals');
-      expect(userMessage.content).toContain('beginner');
-    });
-
-    it('includes notes in user prompt', async () => {
-      const { client, send } = createOpenRouterMockClient();
-      send.mockResolvedValueOnce({
-        choices: [
-          {
-            message: {
-              content: JSON.stringify(VALID_PLAN_RESPONSE),
-            },
-          },
-        ],
-      });
-
-      const provider = new OpenRouterProvider(
-        { model: OPENROUTER_TEST_MODEL },
-        client,
-      );
-      await provider.generate(OPENROUTER_SAMPLE_INPUT);
-
-      const callArgs = send.mock.calls[0][0];
-      const userMessage = callArgs.messages.find(
-        (m: { role: string }) => m.role === 'user',
-      );
-
-      expect(userMessage.content).toContain('Notes: Focus on type safety');
-    });
-
     describe('AsyncIterable streaming path', () => {
       async function* streamEvents(
         payload: string,
@@ -827,182 +470,6 @@ describe('OpenRouterProvider', () => {
         expect(result.metadata.usage?.providerReportedCostUsd).toBeUndefined();
         expect(result.metadata.usage?.promptTokens).toBe(10);
       });
-    });
-  });
-
-  describe('output-token ceiling enforcement', () => {
-    it('sends maxTokens derived from the model ceiling', async () => {
-      const { client, send } = createOpenRouterMockClient();
-      send.mockResolvedValueOnce({
-        choices: [
-          {
-            message: { content: JSON.stringify(VALID_PLAN_RESPONSE) },
-          },
-        ],
-      });
-
-      const provider = new OpenRouterProvider(
-        { model: OPENROUTER_TEST_MODEL },
-        client,
-      );
-      await provider.generate(OPENROUTER_SAMPLE_INPUT);
-
-      const requestBody = send.mock.calls[0][0] as {
-        maxTokens: number;
-        model: string;
-        [key: string]: unknown;
-      };
-      expect(requestBody).toHaveProperty('maxTokens');
-      expect(typeof requestBody.maxTokens).toBe('number');
-      expect(requestBody.maxTokens).toBeGreaterThan(0);
-    });
-
-    it('uses model-specific ceiling for models with explicit maxOutputTokens', async () => {
-      const { client, send } = createOpenRouterMockClient();
-      send.mockResolvedValueOnce({
-        choices: [
-          {
-            message: { content: JSON.stringify(VALID_PLAN_RESPONSE) },
-          },
-        ],
-      });
-
-      // openai/gpt-4o has maxOutputTokens: 64_000
-      const provider = new OpenRouterProvider(
-        { model: 'openai/gpt-4o' },
-        client,
-      );
-      await provider.generate(OPENROUTER_SAMPLE_INPUT);
-
-      const requestBody = send.mock.calls[0][0] as {
-        maxTokens: number;
-        [key: string]: unknown;
-      };
-      expect(requestBody.maxTokens).toBe(
-        getOutputTokenCeiling('openai/gpt-4o'),
-      );
-    });
-
-    it('uses default ceiling for unknown models', async () => {
-      const { client, send } = createOpenRouterMockClient();
-      send.mockResolvedValueOnce({
-        choices: [
-          {
-            message: { content: JSON.stringify(VALID_PLAN_RESPONSE) },
-          },
-        ],
-      });
-
-      const provider = new OpenRouterProvider(
-        { model: 'unknown/model-xyz' },
-        client,
-      );
-      await provider.generate(OPENROUTER_SAMPLE_INPUT);
-
-      const requestBody = send.mock.calls[0][0] as {
-        maxTokens: number;
-        [key: string]: unknown;
-      };
-      expect(requestBody.maxTokens).toBe(DEFAULT_OUTPUT_TOKEN_CEILING);
-    });
-  });
-
-  describe('error handling', () => {
-    it('throws ProviderInvalidResponseError when response is empty', async () => {
-      const { client, send } = createOpenRouterMockClient();
-      send
-        .mockResolvedValueOnce({
-          choices: [],
-        })
-        .mockResolvedValueOnce({
-          choices: [],
-        });
-
-      const provider = new OpenRouterProvider(
-        { model: OPENROUTER_TEST_MODEL },
-        client,
-      );
-
-      await expect(provider.generate(OPENROUTER_SAMPLE_INPUT)).rejects.toThrow(
-        ProviderInvalidResponseError,
-      );
-      await expect(provider.generate(OPENROUTER_SAMPLE_INPUT)).rejects.toThrow(
-        'OpenRouter returned an empty response',
-      );
-    });
-
-    it('throws ProviderInvalidResponseError when content is null', async () => {
-      const { client, send } = createOpenRouterMockClient();
-      send.mockResolvedValueOnce({
-        choices: [
-          {
-            message: {
-              content: null,
-            },
-          },
-        ],
-      });
-
-      const provider = new OpenRouterProvider(
-        { model: OPENROUTER_TEST_MODEL },
-        client,
-      );
-
-      await expect(provider.generate(OPENROUTER_SAMPLE_INPUT)).rejects.toThrow(
-        ProviderInvalidResponseError,
-      );
-    });
-
-    it('passes raw response text through without provider-level schema validation', async () => {
-      const { client, send } = createOpenRouterMockClient();
-      send.mockResolvedValueOnce({
-        choices: [
-          {
-            message: {
-              content: 'not valid json { broken',
-            },
-          },
-        ],
-      });
-
-      const provider = new OpenRouterProvider(
-        { model: OPENROUTER_TEST_MODEL },
-        client,
-      );
-      const result = await provider.generate(OPENROUTER_SAMPLE_INPUT);
-      const rawText = await collectOpenRouterStream(result.stream);
-
-      expect(rawText).toBe('not valid json { broken');
-    });
-
-    it('throws ProviderInvalidResponseError when array content has no text items', async () => {
-      const { client, send } = createOpenRouterMockClient();
-      const noTextContent = {
-        choices: [
-          {
-            message: {
-              content: [
-                { type: 'image', image_url: 'https://example.com/image.png' },
-              ],
-            },
-          },
-        ],
-      };
-      send
-        .mockResolvedValueOnce(noTextContent)
-        .mockResolvedValueOnce(noTextContent);
-
-      const provider = new OpenRouterProvider(
-        { model: OPENROUTER_TEST_MODEL },
-        client,
-      );
-
-      await expect(provider.generate(OPENROUTER_SAMPLE_INPUT)).rejects.toThrow(
-        ProviderInvalidResponseError,
-      );
-      await expect(provider.generate(OPENROUTER_SAMPLE_INPUT)).rejects.toThrow(
-        'no text content',
-      );
     });
   });
 });
